requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0

# Notion integration
notion-client>=2.2.1

# Google Calendar
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# iMessage / macOS integration
pyobjc-framework-Cocoa>=10.0  # For AppleScript execution

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.4

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
